project_manager/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
django-cors-headers==4.7.0
requests==2.32.5
django-tenants==3.8.0
django-environ==0.12.0
django-extensions==3.2.0
django-filter==25.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
PyJWT==2.10.1
cryptography==45.0.6
kafka-python==2.2.15
supabase==2.18.1